+++ R/globals.R
@@ -16,6 +16,7 @@ if (getRversion() >= "2.15.1") {
     # plot_exclusive_junctions
     "transcript_id", "transcript_name", "exon_number",
     "j_start", "j_end", "exclusive", "targeted",
-    "n_tx", "has_excl", "y", "y_label", "xend", "attr", "seqname"
+    "n_tx", "has_excl", "y", "y_label", "xend", "attr", "seqname",
+    "i.start", "i.end"
   ))
 }